refactor: rename namespace from Annotations to Attributes for ResourceFactory
